Create modules folder and move python files
