+++ modules/hash.py
