+++ test_output/krona_dZdXfY9357600.html
@@ -1,0 +1,79 @@
+<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
+<html xmlns="http://www.w3.org/1999/xhtml" xml:lang="en" lang="en">
+ <head>
+  <meta charset="utf-8"/>
+  <link rel="shortcut icon" href="http://krona.sourceforge.net/img/favicon.ico"/>
+  <script id="notfound">window.onload=function(){document.body.innerHTML="Could not get resources from \"http://krona.sourceforge.net\"."}</script>
+  <script src="http://krona.sourceforge.net/src/krona-2.0.js"></script>
+ </head>
+ <body>
+  <img id="hiddenImage" src="http://krona.sourceforge.net/img/hidden.png" style="display:none"/>
+  <img id="loadingImage" src="http://krona.sourceforge.net/img/loading.gif" style="display:none"/>
+  <img id="logo" src="http://krona.sourceforge.net/img/logo.png" style="display:none"/>
+  <noscript>Javascript must be enabled to view this page.</noscript>
+  <div style="display:none">
+  <krona collapse="true" key="true">
+   <attributes magnitude="magnitude">
+    <list>members</list>
+    <attribute display="Total">magnitude</attribute>
+    <attribute display="Unassigned">magnitudeUnassigned</attribute>
+   </attributes>
+   <datasets>
+    <dataset>rpoB_0</dataset>
+   </datasets>
+<node name="all">
+ <magnitude><val>149</val></magnitude>
+ <magnitudeUnassigned><val>16</val></magnitudeUnassigned>
+ <node name="cellular organisms">
+  <magnitude><val>133</val></magnitude>
+  <node name="Eukaryota">
+   <magnitude><val>63</val></magnitude>
+   <node name="Fornicata">
+    <magnitude><val>63</val></magnitude>
+    <node name="Diplomonadida">
+     <magnitude><val>63</val></magnitude>
+     <node name="Hexamitidae">
+      <magnitude><val>63</val></magnitude>
+      <node name="Giardiinae">
+       <magnitude><val>63</val></magnitude>
+       <node name="Giardia">
+        <magnitude><val>63</val></magnitude>
+        <node name="Giardia intestinalis">
+         <magnitude><val>63</val></magnitude>
+        </node>
+       </node>
+      </node>
+     </node>
+    </node>
+   </node>
+  </node>
+  <node name="Bacteria">
+   <magnitude><val>70</val></magnitude>
+   <magnitudeUnassigned><val>7</val></magnitudeUnassigned>
+   <node name="Terrabacteria group">
+    <magnitude><val>63</val></magnitude>
+    <node name="Firmicutes">
+     <magnitude><val>63</val></magnitude>
+     <node name="Bacilli">
+      <magnitude><val>63</val></magnitude>
+      <magnitudeUnassigned><val>61</val></magnitudeUnassigned>
+      <node name="Bacillales">
+       <magnitude><val>2</val></magnitude>
+       <node name="Listeriaceae">
+        <magnitude><val>2</val></magnitude>
+        <node name="Listeria">
+         <magnitude><val>2</val></magnitude>
+         <node name="Listeria monocytogenes">
+          <magnitude><val>2</val></magnitude>
+         </node>
+        </node>
+       </node>
+      </node>
+     </node>
+    </node>
+   </node>
+  </node>
+ </node>
+</node>
+  </krona>
+</div></body></html>
