+++ test_output/normalized_counts_SYhuCb125561317.csv
@@ -1,0 +1,2 @@
+,rpoB_0
+MetagenomeTest_0,1.00000000
