+++ test_output/raw_counts_TtMRFg906042443.csv
@@ -1,0 +1,2 @@
+,rpoB_0
+MetagenomeTest_0,49
