+++ test_output/rpoB_0_refseq_tree_xVeb90649869015.newick
@@ -1,0 +1,1 @@
+(gi|16802304|ref|NP_463789.1|/217-897:0.14136,gi|22536344|ref|NP_687195.1|/220-891:0.23613,(((gi|16802304|ref|NP_463789.1|/7-213:0.13462,gi|22536344|ref|NP_687195.1|/2-220:0.24612)1.000:0.83585,((gi|22536344|ref|NP_687195.1|/893-980:0.09112,(gi|16802304|ref|NP_463789.1|/897-985:0.04848,gi|159116724|ref|XP_001708583.1|/1027-1091:0.97666)0.019:0.05045)1.000:0.39026,gi|16272459|ref|NP_438673.1|/3-1342:0.00084)0.000:0.01291)0.016:0.04327,(((gi|159116724|ref|XP_001708583.1|/1114-1221:1.50051,gi|22536344|ref|NP_687195.1|/979-1184:0.17536)0.439:0.12031,(gi|23465771|ref|NP_696374.1|/13-147:0.01066,(((((gi|23465771|ref|NP_696374.1|/240-625:0.08158,gi|159118719|ref|XP_001709578.1|/301-464:1.43442)0.737:0.29316,gi|159118719|ref|XP_001709578.1|/680-782:1.02422)0.000:0.00054,gi|23465771|ref|NP_696374.1|/1028-1276:0.80092)0.073:0.17515,gi|218562134|ref|YP_002343913.1|/5-1516:0.00871)0.000:0.01179,gi|23465771|ref|NP_696374.1|/656-1029:1.13475)0.099:0.48617)0.982:1.47994)0.065:0.00867,gi|16802304|ref|NP_463789.1|/984-1183:0.09869)0.988:0.41610)0.436:0.46355);
